refactor: rename component files to PascalCase
