+++ src/app/(protected)/dashboard/tarefas/components/AddTaskDialog.tsx
@@ -6,10 +6,10 @@ import { Loader2 } from 'lucide-react'
 import { z } from 'zod'
 import { useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
-import { Button } from '@/components/ui/button'
-import { Input } from '@/components/ui/input'
-import { Textarea } from '@/components/ui/textarea'
-import { DatePicker } from '@/components/ui/date-picker'
+import { Button } from '@/components/ui/Button'
+import { Input } from '@/components/ui/Input'
+import { Textarea } from '@/components/ui/Textarea'
+import { DatePicker } from '@/components/ui/DatePicker'
 import {
   Dialog,
   DialogTrigger,
@@ -17,17 +17,17 @@ import {
   DialogHeader,
   DialogTitle,
   DialogFooter,
-} from '@/components/ui/dialog'
+} from '@/components/ui/Dialog'
 import {
   Select,
   SelectTrigger,
   SelectValue,
   SelectContent,
   SelectItem,
-} from '@/components/ui/select'
+} from '@/components/ui/Select'
 import { priorities } from './data'
 import { createClient } from '@/lib/client'
-import { useNotification } from '@/components/notification-provider'
+import { useNotification } from '@/components/NotificationProvider'
 import {
   Form,
   FormField,
@@ -35,7 +35,7 @@ import {
   FormLabel,
   FormControl,
   FormMessage,
-} from '@/components/ui/form'
+} from '@/components/ui/Form'
 
 const formSchema = z.object({
   title: z.string().min(1, { message: 'Título é obrigatório' }),
