+++ src/app/(protected)/dashboard/tarefas/components/DataTablePagination.tsx
@@ -1,14 +1,14 @@
 import { Table } from "@tanstack/react-table"
 import { ChevronLeft, ChevronRight, ChevronsLeft, ChevronsRight } from "lucide-react"
 
-import { Button } from "@/components/ui/button"
+import { Button } from "@/components/ui/Button"
 import {
   Select,
   SelectContent,
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from "@/components/ui/select"
+} from "@/components/ui/Select"
 
 interface DataTablePaginationProps<TData> {
   table: Table<TData>
