+++ src/app/(protected)/dashboard/tarefas/components/columns.tsx
@@ -2,8 +2,8 @@
 
 import { ColumnDef } from "@tanstack/react-table"
 
-import { Badge } from "@/components/ui/badge"
-import { Checkbox } from "@/components/ui/checkbox"
+import { Badge } from "@/components/ui/Badge"
+import { Checkbox } from "@/components/ui/Checkbox"
 
 import { labels, priorities, statuses } from "./data"
 
@@ -21,8 +21,8 @@ export interface Task {
   associacao: string | null
   tipoId: string | null
 }
-import { DataTableColumnHeader } from "./data-table-column-header"
-import { DataTableRowActions } from "./data-table-row-actions"
+import { DataTableColumnHeader } from "./DataTableColumnHeader"
+import { DataTableRowActions } from "./DataTableRowActions"
 
 export const columns: ColumnDef<Task>[] = [
   {
