+++ src/components/DynamicBreadcrumb.tsx
@@ -9,7 +9,7 @@ import {
   BreadcrumbList,
   BreadcrumbPage,
   BreadcrumbSeparator,
-} from '@/components/ui/breadcrumb'
+} from '@/components/ui/Breadcrumb'
 
 const LABELS: Record<string, string> = {
   dashboard: 'Dashboard',
