+++ src/components/LogoutButton.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { createClient } from '@/lib/client'
-import { Button } from '@/components/ui/button'
+import { Button } from '@/components/ui/Button'
 import { useRouter } from 'next/navigation'
 
 export function LogoutButton() {
