+++ src/components/ui/DatePicker.tsx
@@ -5,9 +5,9 @@ import { format } from "date-fns"
 import { Calendar as CalendarIcon } from "lucide-react"
 
 import { cn } from "@/lib/utils"
-import { Button } from "@/components/ui/button"
-import { Calendar } from "@/components/ui/calendar"
-import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover"
+import { Button } from "@/components/ui/Button"
+import { Calendar } from "@/components/ui/Calendar"
+import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/Popover"
 
 interface DatePickerProps {
   date: Date | undefined
